Linux-specialized the mmap mechanism.
